Work around WebCore failing to build due to NDEBUG getting undefined in release https://bugs.webkit.org/show_bug.cgi?id=225782 <rdar://problem/77982432>
Reviewed by Simon Fraser.

Source/WebCore/PAL:

* pal/spi/cg/CoreGraphicsSPI.h:
* pal/spi/ios/GraphicsServicesSPI.h:
Apply a temporary workaround to preserve NDEBUG through these includes.
There are a few ways to get to CoreGraphicsPrivate.h, so we fix all of them.

Source/WebKit:

* Platform/spi/ios/UIKitSPI.h:
Apply a temporary workaround to preserve NDEBUG through these includes.
There are a few ways to get to CoreGraphicsPrivate.h, so we fix all of them.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277464 268f45cc-cd09-0410-ab3c-d52691b4dbfc