[GTK] Wrong smooth animation sometimes with non-async scrolling https://bugs.webkit.org/show_bug.cgi?id=225764
Reviewed by Adrian Perez de Castro.

This is happening because sometimes when the scrollbars are added to the scroll view, the contents size hasn't
been updated yet, so that values saved by ScrollAnimationSmooth::updateVisibleLengths() are 0x0. After that the
updateVisibleLengths() is not called again when the contents size is updated, so scroll happens with the wrong
visible lengths. This doesn't happen with async scrolling because updateVisibleLengths()  is called every time
layers are repositioned. For non-async scrolling it should be enough to update visible lengths when he contents
size changes.

* platform/ScrollAnimator.cpp:
(WebCore::ScrollAnimator::contentsResized const):
* platform/ScrollAnimator.h:
(WebCore::ScrollAnimator::ScrollAnimator::contentsResized const):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277482 268f45cc-cd09-0410-ab3c-d52691b4dbfc