Bad media session artwork URL results in blue "broken image" icon being shown in Now Playing https://bugs.webkit.org/show_bug.cgi?id=225799 rdar://77990886
Reviewed by Eric Carlson.

Source/WebCore:

By default, when the image's URL doesn't point to an image the CachedImage
will return a default image consistent of a blue interrogation mark.
We need to explicitely check that an error occurred. The naming used in the
method didn't make this need obvious.

* Modules/mediasession/MediaMetadata.cpp:
(WebCore::ArtworkImageLoader::notifyFinished): Check for a status error.
* testing/Internals.cpp:
(WebCore::Internals::loadArtworkImage): Fix spelling

LayoutTests:

* fast/mediasession/metadata/artworkdownload-expected.txt:
* fast/mediasession/metadata/artworkdownload.html: Add test case.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277507 268f45cc-cd09-0410-ab3c-d52691b4dbfc