SHOULD NEVER BE REACHED in FrameSelection::setSelectionWithoutUpdatingAppearance https://bugs.webkit.org/show_bug.cgi?id=225219
Patch by Frederic Wang <fwang@igalia.com> on 2021-05-18
Reviewed by Ryosuke Niwa.

Source/WebCore:

When FrameSelection::selectFrameElementInParentIfFullySelected sets focus on the parent
frame, that can trigger DOM events, possibly making orphan the newSelection prepared before.
This patch fixes that issue by clearing the selection on that parent frame in such a
situation.

Test: editing/selection/selection-in-iframe-removed-assert.html

* editing/FrameSelection.cpp:
(WebCore::FrameSelection::selectFrameElementInParentIfFullySelected): Check if the
newSelection became orphan and if so, clear it.

LayoutTests:

Add a regression test.

* editing/selection/selection-in-iframe-removed-assert.html: Copied from
editing/selection/selection-in-iframe-removed-crash.html, with an additional
requestAnimationFrame.
* editing/selection/selection-in-iframe-removed-assert-expected.txt: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277644 268f45cc-cd09-0410-ab3c-d52691b4dbfc