ASSERTION FAILED: isReactionAllowed() in enqueueDisconnectedCallbackIfNeeded during document teardown https://bugs.webkit.org/show_bug.cgi?id=224033
Reviewed by Maciej Stachowiak.

Source/WebCore:

Moved the bug assertion to after an early exit for when we're in the middle of destorying the document.

enqueueDisconnectedCallbackIfNeeded will be called on custom elements in these circumstances
but there is no correctness issue here since we exit early.


* dom/CustomElementReactionQueue.cpp:
(WebCore::CustomElementReactionQueue::enqueueDisconnectedCallbackIfNeeded):

LayoutTests:

Removed flaky test expectation.

* platform/ios-simulator/TestExpectations:
* platform/mac/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277646 268f45cc-cd09-0410-ab3c-d52691b4dbfc