Accessibility caption styles does not always reflect the selection https://bugs.webkit.org/show_bug.cgi?id=226084 <rdar://75426845>
Reviewed by Chris Fleizach.

Source/WebCore:

Export a soft linked constant.

* WebCore.xcodeproj/project.pbxproj:
* platform/cf/MediaAccessibilitySoftLink.cpp:

Source/WebKit:

This is caused by a race between pushing the caption preference value to the WebContent process and receiving the notification in that process.
This patch addresses this by reposting the notification after having set the preference value in the WebContent process.

* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::captionProfilePreferenceKey):
(WebKit::dispatchSimulatedNotificationsForPreferenceChange):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277874 268f45cc-cd09-0410-ab3c-d52691b4dbfc