Adopt CheckedLock in RemoteImageBufferProxy https://bugs.webkit.org/show_bug.cgi?id=226098
Reviewed by Tim Horton.

Adopt CheckedLock in RemoteImageBufferProxy to benefit from Clang Thread Safety Analysis.
I had to use WTF_IGNORES_THREAD_SAFETY_ANALYSIS on one of the getter where we weren't
locking. My understanding is that the code is currently safe without the lock in this
case and I added a comment and assertions to clarify that.

* WebProcess/GPU/graphics/RemoteImageBufferProxy.h:
(WebKit::RemoteImageBufferProxy::waitForDidFlushOnSecondaryThread):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277892 268f45cc-cd09-0410-ab3c-d52691b4dbfc