CrashOnOverflow in WebCore::RenderTable::cellBefore(WebCore::RenderTableCell const*) https://bugs.webkit.org/show_bug.cgi?id=225949
Reviewed by Sergio Villar Senin.

Source/WebCore:

The table section maintains a grid of rows and columns. This grid is used to jump to cells directly during layout (as opposed to tree walking).
It needs to be re-populated on tree mutations e.g when a new row is added to the table.
RenderTableSection::willInsertTableRow assumes that rows and cells are inserted subsequently, after each other and
it's ok to ignore the grid invalidation on the first row insertion (beforeChild is nullptr).
However in case of continuation split (e.g <table> is inside of an inline element and this inline element
needs to be split due to a block level box insertion), the row and all of its cells are cloned and inserted under the table section at once.

Test: fast/table/table-split-on-insert-with-section-crash.html

* rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::willInsertTableRow): fun fact: this code dates back to r3351 "The new table code from Lars.".

LayoutTests:

* fast/table/table-split-on-insert-with-section-crash-expected.txt: Added.
* fast/table/table-split-on-insert-with-section-crash.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278028 268f45cc-cd09-0410-ab3c-d52691b4dbfc