REGRESSION (r244434) Dispatch intrinsicContentsSizeChanged on the main frame only https://bugs.webkit.org/show_bug.cgi?id=226311
Reviewed by Tim Horton.

Source/WebCore:

While the Mail version of the content autosizing runs on both the main frame and the subframe(s),
the content size reporting should only happen on the main frame. The client is only interested
in the enclosing size of the content (also the API to retrieve the information is per-WKWebView,
it doesn't expose this information on a frame-by-frame basis).

Test: fast/dynamic/mail-autosize-with-iframe.html

* page/FrameView.cpp:
(WebCore::FrameView::autoSizeIfEnabled):
(WebCore::FrameView::enableFixedWidthAutoSizeMode):
(WebCore::FrameView::enableSizeToContentAutoSizeMode):
(WebCore::FrameView::enableAutoSizeMode):

LayoutTests:

* fast/dynamic/mail-autosize-with-iframe-expected.txt: Added.
* fast/dynamic/mail-autosize-with-iframe.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@278164 268f45cc-cd09-0410-ab3c-d52691b4dbfc